backend/pyproject.toml
[project]
name = "axiome-backend"
version = "0.1.0"
description = "Python backend for Axiome algorithm learning platform"
authors = [
    {name = "Suhas Prakash", email = "suh4s@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["axiome_backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.9"
strict = false
warn_return_any = true
warn_unused_configs = true
